+++ examples/1-getting-started/src/server/index.js
@@ -1,6 +1,5 @@
-const { ApolloServer, graphiqlExpress } = require('apollo-server-express');
+const { ApolloServer } = require('apollo-server-express');
 const { createServer } = require('http');
-const { SubscriptionServer } = require('subscriptions-transport-ws');
 const expressPlayground = require('graphql-playground-middleware-express')
   .default;
 const express = require('express');
@@ -22,9 +21,7 @@ const webServer = createServer(app);
 server.installSubscriptionHandlers(webServer);
 
 const graphqlEndpoint = `http://localhost:${PORT}${server.graphqlPath}`;
-const subscriptionEndpoint = `ws://localhost:${PORT}${
-  server.subscriptionsPath
-}`;
+const subscriptionEndpoint = `ws://localhost:${PORT}${server.subscriptionsPath}`;
 
 app.use(cors());
 app.get(