+++ src/index.ts
@@ -7,6 +7,7 @@ export * from './types';
 
 export {
   CombinedError,
+  stringifyVariables,
   createRequest,
   makeResult,
   makeErrorResult,