+++ src/utils/index.ts
@@ -3,6 +3,8 @@ export * from './request';
 export * from './result';
 export * from './typenames';
 export * from './toSuspenseSource';
+export * from './stringifyVariables';
+export * from './withPromise';
 
 export const noop = () => {
   /* noop */